% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exact.R
\name{make_flat_log_prior}
\alias{make_flat_log_prior}
\title{Make a flat prior for maximum likelihood estimation}
\usage{
make_flat_log_prior(...)
}
\arguments{
\item{...}{Currently not used}
}
\value{
A `prior` object with a constant log density and a log gradient equal to zero
}
\description{
Make a flat prior for maximum likelihood estimation
}

