% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{make_flat_prior}
\alias{make_flat_prior}
\title{Make a flat prior for maximum likelihood estimation}
\usage{
make_flat_prior(...)
}
\arguments{
\item{...}{Currently not used}
}
\value{
A \code{prior} object, which is a list of two functions:
\item{log_d}{A function for calculating the log of the prior density for each element of a vector. 
With a uniform prior, the log-density is always zero.}
\item{log_grad}{A function for calculating the gradient of the log-density for each element of 
a vector. For the uniform prior, this is always 0. With a uniform prior, this gradient is always zero.}
}
\description{
Make a flat prior for maximum likelihood estimation
}
\examples{
p = make_flat_prior()
curve(p$log_d(x), from = -5, to = 5)
curve(p$log_grad(x), from = -5, to = 5)
}
\seealso{
\code{\link{make_logistic_prior}}
}

