% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exact.R
\name{rosalia}
\alias{rosalia}
\title{Fit a Markov network to binary data}
\usage{
rosalia(x, prior = make_flat_prior(), maxit = 100, trace = 1,
  hessian = FALSE, parlist, ...)
}
\arguments{
\item{x}{A binary matrix.}

\item{prior}{An object of class \code{prior}. By default, the prior is flat for maximum likelihood estimation}

\item{maxit,trace,hessian,...}{Arguments passed to \code{\link{optim}}}

\item{parlist}{user-specified starting values (optional).}
}
\description{
Fit a Markov network to binary data
}

