% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exact.R
\name{make_logistic_prior}
\alias{make_logistic_prior}
\title{Make a logistic prior with specified scale}
\usage{
make_logistic_prior(location = 0, scale, ...)
}
\arguments{
\item{location}{Numeric vector; see \code{\link{Logistic}}}

\item{scale}{Numeric vector; see \code{\link{Logistic}}}

\item{...}{Currently not used}
}
\value{
A `prior` object with a log density and a log gradient
}
\description{
Make a logistic prior with specified scale
}

