% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exact.R
\name{rosalia}
\alias{rosalia}
\title{Fit a Markov network to binary data}
\usage{
rosalia(x, prior = make_flat_prior(), maxit = 100, trace = 1,
  hessian = FALSE, parlist, ...)
}
\arguments{
\item{x}{A binary matrix.}

\item{prior}{An object of class \code{prior}. By default, the prior is flat for maximum likelihood estimation}

\item{maxit,trace,hessian,...}{Arguments passed to \code{\link{optim}}}

\item{parlist}{user-specified starting values (optional).}
}
\value{
a \code{list} with the following elements:
\item{alpha}{A vector of estimated univariate potentials ("intercepts"). Larger values
support higher probabilities for the corresponding node.}
\item{beta}{A symmetric matrix of bivariate potentials ("interaction strengths"). Larger values
support higher probabilities for the corresponding pair of nodes.}
\item{prior}{The prior object that was used during model fitting.}
\item{opt}{The list returned by \code{\link[stats]{optim}}.}
}
\description{
Fit a Markov network to binary data
}
\examples{
# Simulate a random binary matrix with 1000 observations of five binary variables
m = matrix(rbinom(5000, size = 1, prob = .5), nrow = 1000, ncol = 5)

fit = rosalia(m)

fit$alpha
fit$beta
}

